requirements.txt
streamlit
yt-dlp
google-generativeai
fpdf2
